moved forms to component folder
