+++ frontend/app/src/bots/body.tsx
@@ -10,8 +10,8 @@ import { colors } from '../colors';
 import FadeLeft from '../animated/fadeLeft';
 import { useIsMobile } from '../hooks';
 import Bot from './bot';
-import Form from '../form';
-import { botSchema } from '../form/schema';
+import Form from '../components/form';
+import { botSchema } from '../components/form/schema';
 import MaterialIcon from '@material/react-material-icon';
 import BotView from './botView';
 import BotSecret from './utils/botSecret';
