+++ frontend/app/src/components/form/inputs/creatable-multi-select-input-new.tsx
@@ -2,14 +2,14 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { EuiCheckboxGroup, EuiPopover, EuiText } from '@elastic/eui';
 import type { Props } from './propsType';
-import { colors } from '../../colors';
-import ImageButton from '../../sphinxUI/Image_button';
+import { colors } from '../../../colors';
+import ImageButton from '../../../sphinxUI/Image_button';
 import {
   coding_languages,
   GetValue,
   LanguageObject
-} from '../../people/utils/language_label_style';
-import { SvgMask } from '../../people/utils/svgMask';
+} from '../../../people/utils/language_label_style';
+import { SvgMask } from '../../../people/utils/svgMask';
 
 const codingLanguages = GetValue(coding_languages);
