+++ frontend/app/src/components/form/inputs/gallery-input.tsx
@@ -2,12 +2,12 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import Dropzone from 'react-dropzone';
 import type { Props } from './propsType';
-import { useStores } from '../../store';
+import { useStores } from '../../../store';
 import MaterialIcon from '@material/react-material-icon';
 
-import { Button, Modal } from '../../sphinxUI';
-import { MAX_UPLOAD_SIZE } from '../../people/utils/constants';
-import { colors } from '../../colors';
+import { Button, Modal } from '../../../sphinxUI';
+import { MAX_UPLOAD_SIZE } from '../../../people/utils/constants';
+import { colors } from '../../../colors';
 
 export default function GalleryInput({
   value,
