+++ frontend/app/src/components/form/inputs/loom-video-input-new.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import type { Props } from './propsType';
-import { colors } from '../../colors';
+import { colors } from '../../../colors';
 import { EuiText } from '@elastic/eui';
-import LoomViewerRecorderNew from '../../people/utils/loomViewerRecorderNew';
+import LoomViewerRecorderNew from '../../../people/utils/loomViewerRecorderNew';
 
 export default function LoomVideoInputNew({
   value,
