+++ frontend/app/src/components/form/inputs/multi-select-input.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components';
 import { EuiIcon } from '@elastic/eui';
 import type { Props } from './propsType';
 import { FieldEnv, Note } from './index';
-import { MultiSelect } from '../../sphinxUI';
-import { colors } from '../../colors';
+import { MultiSelect } from '../../../sphinxUI';
+import { colors } from '../../../colors';
 
 export default function MultiSelectInput({
   error,
