+++ frontend/app/src/components/form/inputs/searchable-select-input.tsx
@@ -3,9 +3,9 @@ import styled from 'styled-components';
 import { EuiIcon } from '@elastic/eui';
 import type { Props } from './propsType';
 import { FieldEnv, Note } from './index';
-import { SearchableSelect } from '../../sphinxUI';
-import { useStores } from '../../store';
-import { colors } from '../../colors';
+import { SearchableSelect } from '../../../sphinxUI';
+import { useStores } from '../../../store';
+import { colors } from '../../../colors';
 
 export default function SearchableSelectInput({
   error,
