+++ frontend/app/src/components/form/inputs/text-area-input.tsx
@@ -1,10 +1,22 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
+import { EuiIcon } from '@elastic/eui';
 import type { Props } from './propsType';
-import { FieldEnv, FieldText, Note } from './index';
-import { colors } from '../../colors';
+import { FieldEnv, FieldTextArea, Note } from './index';
+import { colors } from '../../../colors';
 
-export default function TextInput({
+const StyleOnText = {
+  Description: {
+    height: '172px',
+    width: '292px'
+  },
+  Deliverables: {
+    height: '135px',
+    width: '192px'
+  }
+};
+
+export default function TextAreaInput({
   error,
   note,
   label,
@@ -13,16 +25,13 @@ export default function TextInput({
   handleBlur,
   handleFocus,
   readOnly,
-  prepend,
   extraHTML,
   borderType
 }: Props) {
   let labeltext = label;
+  const color = colors['light'];
   if (error) labeltext = `${labeltext} (${error})`;
   const [active, setActive] = useState<boolean>(false);
-  const color = colors['light'];
-
-  const padStyle = prepend ? { paddingLeft: 0 } : {};
   return (
     <OuterContainer color={color}>
       <FieldEnv
@@ -33,13 +42,15 @@ export default function TextInput({
         className={active ? 'euiFormRow_active' : (value ?? '') === '' ? '' : 'euiFormRow_filed'}
         border={borderType}
         label={labeltext}
-        isTextField={true}
-        error={error}
+        height={StyleOnText[label].height}
+        width={StyleOnText[label].width}
       >
         <R>
-          <FieldText
+          <FieldTextArea
             color={color}
-            name={'first'}
+            height={StyleOnText[label].height}
+            width={StyleOnText[label].width}
+            name="first"
             value={value || ''}
             readOnly={readOnly || false}
             onChange={(e) => handleChange(e.target.value)}
@@ -51,12 +62,12 @@ export default function TextInput({
               handleFocus(e);
               setActive(true);
             }}
-            prepend={prepend}
-            style={padStyle}
-            isTextField={true}
+            rows={label === 'Description' ? 8 : 6}
           />
           {error && (
-            <E color={color} />
+            <E color={color}>
+              <EuiIcon type="alert" size="m" style={{ width: 20, height: 20 }} />
+            </E>
           )}
         </R>
       </FieldEnv>
@@ -76,15 +87,15 @@ interface styledProps {
 
 const OuterContainer = styled.div<styledProps>`
   .euiFormRow_active {
-    border: 1px solid ${(p) => p.color && p.color.blue2};
+    border: 1px solid ${(p) => p?.color && p?.color.blue2};
     .euiFormRow__labelWrapper {
       margin-bottom: 0px;
       margin-top: -9px;
       padding-left: 10px;
       height: 14px;
       label {
-        color: ${(p) => p.color && p.color.grayish.G300} !important;
-        background: ${(p) => p.color && p.color.pureWhite};
+        color: ${(p) => p?.color && p?.color.grayish.G300} !important;
+        background: ${(p) => p?.color && p?.color.pureWhite};
         z-index: 10;
       }
     }
@@ -96,8 +107,8 @@ const OuterContainer = styled.div<styledProps>`
       padding-left: 10px;
       height: 14px;
       label {
-        color: ${(p) => p.color && p.color.grayish.G300} !important;
-        background: ${(p) => p.color && p.color.pureWhite};
+        color: ${(p) => p?.color && p?.color.grayish.G300} !important;
+        background: ${(p) => p?.color && p?.color.pureWhite};
         z-index: 10;
       }
     }
@@ -105,10 +116,8 @@ const OuterContainer = styled.div<styledProps>`
 `;
 
 const ExtraText = styled.div<styledProps>`
-  padding: 0px 10px 5px;
-  margin: -5px 0 10px;
-  color: ${(p) => p.color && p.color.red3};
-  font-style: italic;
+  color: ${(p) => p?.color && p?.color.grayish.G760};
+  padding: 10px 10px 25px 10px;
   max-width: calc(100% - 20px);
   word-break: break-all;
   font-size: 14px;
@@ -117,12 +126,11 @@ const ExtraText = styled.div<styledProps>`
 const E = styled.div<styledProps>`
   position: absolute;
   right: 10px;
-  top: 0px;
+  top: 10px;
   display: flex;
-  height: 100%;
   justify-content: center;
   align-items: center;
-  color: ${(p) => p.color && p.color.blue3};
+  color: ${(p) => p?.color && p?.color.blue3};
   pointer-events: none;
   user-select: none;
 `;
