+++ frontend/app/src/components/form/inputs/text-input-new.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import type { Props } from './propsType';
-import { colors } from '../../colors';
+import { colors } from '../../../colors';
 
-export default function TextAreaInputNew({
+export default function TextInputNew({
   error,
   label,
   value,
@@ -26,9 +26,10 @@ export default function TextAreaInputNew({
 
   return (
     <InputOuterBox color={color} borderColor={isError ? color.red2 : color.grayish.G600}>
-      <textarea
+      <input
         className="inputText"
         id={'text'}
+        type={'text'}
         value={textValue}
         onFocus={handleFocus}
         onBlur={() => {
@@ -41,9 +42,6 @@ export default function TextAreaInputNew({
           handleChange(e.target.value);
           setTextValue(e.target.value);
         }}
-        style={{
-          height: label === 'Deliverables' ? '137px' : '175px'
-        }}
       />
       <label
         htmlFor={'text'}
@@ -52,7 +50,7 @@ export default function TextAreaInputNew({
         style={{
           position: 'absolute',
           left: 16,
-          top: !isFocused[label] ? (textValue === undefined || textValue === '' ? 10 : -9) : -9,
+          top: !isFocused[label] ? (textValue === undefined ? 10 : -9) : -9,
           fontSize: !isFocused[label] ? (textValue === undefined ? 14 : 12) : 12,
           color: color.grayish.G300,
           background: color.pureWhite,
@@ -74,8 +72,9 @@ interface styledProps {
 
 const InputOuterBox = styled.div<styledProps>`
   position: relative;
-  margin-bottom: 0px;
+  margin-bottom: 32px;
   .inputText {
+    height: 40px;
     width: 292px;
     font-size: 14px;
     color: ${(p) => p.color && p.color.pureBlack};
@@ -83,9 +82,7 @@ const InputOuterBox = styled.div<styledProps>`
     border-radius: 4px;
     outline: none;
     padding-left: 16px;
-    padding-top: 16px;
-    resize: none;
-    color: #3c3f41;
+    color: ${(p) => p.color && p.color.text2};
     font-weight: 500;
     letter-spacing: 0.01em;
     :active {
