+++ frontend/app/src/components/form/inputs/text-input.tsx
@@ -1,22 +1,10 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { EuiIcon } from '@elastic/eui';
 import type { Props } from './propsType';
-import { FieldEnv, FieldTextArea, Note } from './index';
-import { colors } from '../../colors';
+import { FieldEnv, FieldText, Note } from './index';
+import { colors } from '../../../colors';
 
-const StyleOnText = {
-  Description: {
-    height: '172px',
-    width: '292px'
-  },
-  Deliverables: {
-    height: '135px',
-    width: '192px'
-  }
-};
-
-export default function TextAreaInput({
+export default function TextInput({
   error,
   note,
   label,
@@ -25,13 +13,16 @@ export default function TextAreaInput({
   handleBlur,
   handleFocus,
   readOnly,
+  prepend,
   extraHTML,
   borderType
 }: Props) {
   let labeltext = label;
-  const color = colors['light'];
   if (error) labeltext = `${labeltext} (${error})`;
   const [active, setActive] = useState<boolean>(false);
+  const color = colors['light'];
+
+  const padStyle = prepend ? { paddingLeft: 0 } : {};
   return (
     <OuterContainer color={color}>
       <FieldEnv
@@ -42,15 +33,13 @@ export default function TextAreaInput({
         className={active ? 'euiFormRow_active' : (value ?? '') === '' ? '' : 'euiFormRow_filed'}
         border={borderType}
         label={labeltext}
-        height={StyleOnText[label].height}
-        width={StyleOnText[label].width}
+        isTextField={true}
+        error={error}
       >
         <R>
-          <FieldTextArea
+          <FieldText
             color={color}
-            height={StyleOnText[label].height}
-            width={StyleOnText[label].width}
-            name="first"
+            name={'first'}
             value={value || ''}
             readOnly={readOnly || false}
             onChange={(e) => handleChange(e.target.value)}
@@ -62,12 +51,12 @@ export default function TextAreaInput({
               handleFocus(e);
               setActive(true);
             }}
-            rows={label === 'Description' ? 8 : 6}
+            prepend={prepend}
+            style={padStyle}
+            isTextField={true}
           />
           {error && (
-            <E color={color}>
-              <EuiIcon type="alert" size="m" style={{ width: 20, height: 20 }} />
-            </E>
+            <E color={color} />
           )}
         </R>
       </FieldEnv>
@@ -87,15 +76,15 @@ interface styledProps {
 
 const OuterContainer = styled.div<styledProps>`
   .euiFormRow_active {
-    border: 1px solid ${(p) => p?.color && p?.color.blue2};
+    border: 1px solid ${(p) => p.color && p.color.blue2};
     .euiFormRow__labelWrapper {
       margin-bottom: 0px;
       margin-top: -9px;
       padding-left: 10px;
       height: 14px;
       label {
-        color: ${(p) => p?.color && p?.color.grayish.G300} !important;
-        background: ${(p) => p?.color && p?.color.pureWhite};
+        color: ${(p) => p.color && p.color.grayish.G300} !important;
+        background: ${(p) => p.color && p.color.pureWhite};
         z-index: 10;
       }
     }
@@ -107,8 +96,8 @@ const OuterContainer = styled.div<styledProps>`
       padding-left: 10px;
       height: 14px;
       label {
-        color: ${(p) => p?.color && p?.color.grayish.G300} !important;
-        background: ${(p) => p?.color && p?.color.pureWhite};
+        color: ${(p) => p.color && p.color.grayish.G300} !important;
+        background: ${(p) => p.color && p.color.pureWhite};
         z-index: 10;
       }
     }
@@ -116,8 +105,10 @@ const OuterContainer = styled.div<styledProps>`
 `;
 
 const ExtraText = styled.div<styledProps>`
-  color: ${(p) => p?.color && p?.color.grayish.G760};
-  padding: 10px 10px 25px 10px;
+  padding: 0px 10px 5px;
+  margin: -5px 0 10px;
+  color: ${(p) => p.color && p.color.red3};
+  font-style: italic;
   max-width: calc(100% - 20px);
   word-break: break-all;
   font-size: 14px;
@@ -126,11 +117,12 @@ const ExtraText = styled.div<styledProps>`
 const E = styled.div<styledProps>`
   position: absolute;
   right: 10px;
-  top: 10px;
+  top: 0px;
   display: flex;
+  height: 100%;
   justify-content: center;
   align-items: center;
-  color: ${(p) => p?.color && p?.color.blue3};
+  color: ${(p) => p.color && p.color.blue3};
   pointer-events: none;
   user-select: none;
 `;
