+++ frontend/app/src/components/form/inputs/widgets/focusedWidget.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import Input from '../../../form/inputs/';
+import Input from '..';
 import { EuiButton } from '@elastic/eui';
 import WidgetList from './widgetList';
 import MaterialIcon from '@material/react-material-icon';
