+++ frontend/app/src/components/form/inputs/widgets/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import Widget from './widget';
 import FocusedWidget from './focusedWidget';
-import FadeLeft from '../../../animated/fadeLeft';
+import FadeLeft from '../../../../animated/fadeLeft';
 import styled from 'styled-components';
 import { useEffect } from 'react';
