+++ frontend/app/src/components/form/inputs/widgets/interfaces.ts
