+++ frontend/app/src/components/form/inputs/widgets/listItems/offer.tsx
