+++ frontend/app/src/people/utils/constants.tsx
@@ -8,7 +8,7 @@ import {
   offerOtherSchema,
   wantedCodingTaskSchema,
   wantedOtherSchema
-} from '../../form/schema';
+} from '../../components/form/schema';
 
 const MAX_UPLOAD_SIZE = 10194304; //10MB
