+++ frontend/app/src/people/widgetViews/summaries/wantedSummary.tsx
@@ -5,9 +5,9 @@ import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import api from '../../../api';
 import { colors } from '../../../colors';
-import Form from '../../../form';
-import InvitePeopleSearch from '../../../form/inputs/widgets/PeopleSearch';
-import { sendBadgeSchema } from '../../../form/schema';
+import Form from '../../../components/form';
+import InvitePeopleSearch from '../../../components/form/inputs/widgets/PeopleSearch';
+import { sendBadgeSchema } from '../../../components/form/schema';
 import { extractGithubIssue, extractGithubIssueFromUrl, formatPrice, satToUsd } from '../../../helpers';
 import { useIsMobile } from '../../../hooks';
 import { Button, Divider, Modal, Paragraph, Title } from '../../../sphinxUI';
